Reverse dims=:lat, :lon for LowerTriangularArray
