+++ src/LowerTriangularMatrices/LowerTriangularMatrices.jl
@@ -19,7 +19,7 @@ export eachharmonic, eachmatrix
 export OneBased, ZeroBased
 
 include("lower_triangular_array.jl")
-include("rotate.jl")
+include("rotate_reverse.jl")
 include("plot.jl")
 
 end
